… CHANGELOG.md
… sdk/metric/EXPERIMENTAL.md
… sdk/metric/internal/aggregate/aggregate_test.go
… sdk/metric/internal/aggregate/exponential_histogram_test.go
… sdk/metric/internal/aggregate/histogram_test.go