@MrAlias 2024-01-30
> implementation looks good, but it looks like it needs unit tests?

Right. Added https://github.com/open-telemetry/opentelemetry-go/pull/4871/commits/c589d01f0443a9d86bdefd7b3886751c3b18c0fc

PTAL